requirements.txt
datadog-api-client
python-dotenv
requests
